arcli publish function complete.

+++ commands/package/publish/actions.js
@@ -1,4 +1,5 @@
 const op = require('object-path');
+const crypto = require('crypto');
 const _ = require('underscore');
 const chalk = require('chalk');
 const fs = require('fs-extra');
@@ -6,7 +7,15 @@ const path = require('path');
 const tar = require('tar');
 
 module.exports = spinner => {
-    let bytes, cwd, filename, pkg, pkgFile, pkgUpdate, prompt, sessionToken;
+    let bytes,
+        cwd,
+        filename,
+        filepath,
+        pkg,
+        pkgFile,
+        pkgUpdate,
+        prompt,
+        sessionToken;
     const x = chalk.magenta('✖');
 
     const message = text => {
@@ -24,8 +33,18 @@ module.exports = spinner => {
         new Promise((resolve, reject) => {
             const required = true;
             const type = 'string';
+            const defaults = {
+                reactium: {
+                    version: '3.2.6',
+                },
+                version: '0.0.1',
+            };
             const schema = fields.map(({ description, name }) => ({
+                default: op.get(defaults, name),
                 description,
+                message: `${chalk.cyan.bold(name)} ${chalk.white(
+                    'is required',
+                )}`,
                 name,
                 required,
                 type,
@@ -59,12 +78,16 @@ module.exports = spinner => {
             pkgFile = path.normalize(`${cwd}/package.json`);
             prompt = props.prompt;
             sessionToken = op.get(props, 'config.registry.sessionToken');
+
+            const app = op.get(props, 'config.registry.app', 'Actinium');
+            const serverURL = op.get(props, 'config.registry.server');
+            Actinium.initialize(app);
+            Actinium.serverURL = serverURL;
         },
         package: async ({ action, params, props }) => {
-            message(`checking ${chalk.cyan('package.json')}...`);
+            message(`updating ${chalk.cyan('package.json')}...`);
             const fields = {
                 name: chalk.white('Package Name:'),
-                version: chalk.white('Version:'),
                 description: chalk.white('Description:'),
                 author: chalk.white('Author:'),
                 'reactium.version': chalk.white('Reactium Version:'),
@@ -90,6 +113,7 @@ module.exports = spinner => {
 
                 pkgUpdate = true;
                 spinner.stopAndPersist({ symbol: x, text });
+
                 console.log('');
                 await pkgPrompt(
                     Object.entries(fields).map(([name, description]) => ({
@@ -98,6 +122,7 @@ module.exports = spinner => {
                     })),
                 );
                 console.log('');
+
                 spinner.start();
             } else {
                 // get package.json file
@@ -135,28 +160,52 @@ module.exports = spinner => {
             }
 
             // Write new package.json
-            if (!pkgUpdate) return;
+            pkg.version = params.version;
+
+            let pname = String(pkg.name)
+                .toLowerCase()
+                .replace(/\s\s+/g, ' ')
+                .replace(/ /g, '-')
+                .replace(/[^0-9a-z@\-\/]/gi, '');
+
+            pname = pname.startsWith('@reactium-module/')
+                ? pname
+                : `@reactium-module/${pname}`;
+
+            op.set(pkg, 'name', pname);
             fs.writeFileSync(pkgFile, JSON.stringify(pkg, null, 2));
+            spinner.start();
+        },
+        validate: async ({ action, params, props }) => {
+            const { name, version } = pkg;
+
+            const canPublish = await Actinium.Cloud.run(
+                'registry-check',
+                { name, version },
+                { sessionToken },
+            );
+
+            if (canPublish !== true) {
+                spinner.fail(`${chalk.magenta('Error:')} ${canPublish}`);
+                exit();
+            }
         },
         compress: ({ action, params, props }) => {
-            //message(`packaging ${chalk.cyan(path.basename(cwd))}...`);
             spinner.stopAndPersist({
                 symbol: chalk.cyan('+'),
-                text: `packaging ${chalk.cyan(path.basename(cwd))}...`,
+                text: `packaging ${chalk.cyan(pkg.name)}...`,
             });
 
             filename = ['reactium-module', 'tgz'].join('.');
-            const filepath = path.normalize(path.join(cwd, filename));
+            filepath = path.normalize(path.join(cwd, filename));
 
-            if (fs.existsSync(filepath)) {
-                fs.removeSync(filepath);
-            }
+            if (fs.existsSync(filepath)) fs.removeSync(filepath);
 
             bytes = 0;
 
-            let dir = '/'+path.basename(cwd)+'/';
+            const dir = '/' + path.basename(cwd) + '/';
 
-            return tar.create(
+            tar.create(
                 {
                     file: filepath,
                     filter: (file, stat) => {
@@ -170,7 +219,7 @@ module.exports = spinner => {
                                 '   ',
                                 bytesToSize(stat.size),
                                 chalk.cyan('→'),
-                                file.split(dir)[0],
+                                dir + file,
                             );
                         }
                         return true;
@@ -180,14 +229,48 @@ module.exports = spinner => {
                 },
                 ['./', filename],
             );
-        },
-        upload: ({ action, params, props }) => {
 
-            console.log(chalk.cyan('+'), 'packaged', bytesToSize(bytes));
+            const { size } = fs.statSync(filepath);
+
+            console.log(chalk.cyan('+'), 'compressed', bytesToSize(bytes), '→', bytesToSize(size));
             console.log('');
-
             spinner.start();
-            message(`uploading ${chalk.cyan(filename)}...`);
         },
+        publish: async ({ action, params, props }) => {
+            message(`processing ${chalk.cyan(filename)}...`);
+
+            let filedata = fs.readFileSync(filepath);
+
+            const checksum = crypto
+                .createHash('sha256')
+                .update(filedata)
+                .digest('hex');
+
+            message(`uploading ${chalk.cyan(filename)}...`);
+            filedata = Array.from(filedata);
+            const file = await new Actinium.File(`reactium-module.${checksum}.tgz`, filedata).save();
+
+            message(`publishing ${chalk.cyan(filename)}...`);
+
+            const data = {
+                checksum,
+                file: file.url(),
+                name: pkg.name,
+                private: params.private,
+                version: String(pkg.version),
+            };
+
+            const result = await Actinium.Cloud.run('registry-publish', data, {
+                sessionToken,
+            });
+
+            spinner.stopAndPersist({
+                symbol: chalk.green('✔'),
+                text: `published ${chalk.cyan(pkg.name)} v${chalk.cyan(pkg.version)}`
+            });
+        },
+        cleanup: () => {
+            fs.removeSync(filepath);
+        }
     };
 };
